==> vignettes/1-select-data.Rmd <==
---
title: "Seleção do período de análise dados das EMA do INMET"
author: "Jônatan Tatsch"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_notebook:
    toc: yes
  html_document:
    fig_caption: yes
    fig_width: 6
    keep_md: yes
    number_sections: yes
    toc: yes

---


# Introdução 

A definição do período para análise das séries de dados de uma rede de estações meteorológicas automáticas (EMAs) segue critérios baseados na homogeneidade espacial das EMAs, na quantidade de dados mínima. Como critério inicial serão selecionadas estações meteorológicas automáticas (EMA) com pelo menos 4 anos de dados (podem ser descontínuos).

As etapas para seleção do período são:

- regularizar as séries temporais de cada EMA para garantir que todas tenham 24 h em cada dia e 365 (ou 366 dias, se ano bissexto) em cada ano

- determinar o período de dados de cada EMA (n° de anos)

- quantificar o número de dados válidos (não faltantes) para cada EMA


# Pacotes, funções e pré-configurações

Carregando pacotes necessários.

```{r setup, message=FALSE}
# limpa espaço de trabalho
rm(list = ls()) 
# carrega pacotes necessários
pacotes <- c("knitr", "tidyverse", "lubridate", "openair", "stringr", 
             "magrittr", "padr")
easypackages::libraries(pacotes)
# configura chuncks
# carrega scripts necessários
source("../R/gaps.R")
source("../R/utils.R")
source("../R/gg_bubble.R")
source("../R/complete_dates.R")
```

Os dados são armazenados usando o Tempo Universal Coordenado [UTC](https://pt.wikipedia.org/wiki/Tempo_Universal_Coordenado). Portanto, podemos definir esse horário como padrão nessa sessão do R. Assim qualquer conversão ou operação com datas e horários usando objetos [POSIX](https://stat.ethz.ch/R-manual/R-devel/library/base/html/DateTimeClasses.html) serão realizadas em UTC.

```{r chunk1}
# definindo globalmente tz = "UTC"
Sys.setenv(TZ = "UTC")
```

# Dados brutos

## Informações das EMAs (metadados)

Informações como as coordenadas de localização, altitude, nome e código das EMAs foram extraídas do cabeçalho contido no arquivo excel com os dados de cada EMA. Entretanto, notou-se que em alguns casos houve erro de digitação das coordenadas quando comparadas aquelas disponíveis no site do INMET.

```{r chunk4}
# metadados do INMET corrigidos e atualizados.(see R/aws-inmet-metadata.R)
tar_info <- readRDS("../data/info-inmet-sul.rds")
  tar_info
```


## Dados meteorológicos

```{r}
# carregar dados meteorológicos das EMA
tar_data <- readRDS("../data/data-raw-inmet-sul-20000922-20161231.rds") %>%
  select(site:tair, tmin, tmax, prec, p, rg)
glimpse(tar_data)
#"A825" %in% unique(tar_data$site)
```


# Dados temporalmente consistentes

Os dados podem conter falhas temporais. Isso implica que podem haver saltos temporais nos dados. Para garantir que os dados fiquem consitentes temporal (cada estação com dias contendo 24 horas e anos com 365 ou 366 dias) faz-se o preenchimento dos horários faltantes com as datas adequadas e os valores das variáveis recebem `NA`.

```{r, eval=TRUE}
# verificando intervalo de tempo 
time_step <- tar_data %>%
  filter(., site == "A801") %>%
  select(date) %$%
  get_interval(date)
time_step
nrow(tar_data)
```

```{r, include = FALSE}
nrow(tar_data)
nrow(distinct(tar_data))
```


```{r}
# cada site com um perído de dados diferente
table(tar_data$site)
#microbenchmark::microbenchmark(
tar_data_pad <- tar_data %>%
  filter(!is.na(date)) %>%
   ## pad é muito lento
   # pad(interval = time_step, 
   #     group = "site", 
   #     break_above = 20
   #     #start_val = min(tar_data$date), 
   #     #end_val = max(tar_data$date)
   #     ) #%>%
  complete_dates(group = "site", time_step = "hours")
#, times = 1)
nrow(tar_data_pad)
# verificaçao da padronizaçao temporal
table(tar_data_pad$site)
#range(tar_data_pad2$date)
```

```{r, include=FALSE}
table(tar_data$site)
tar_data_pad2 <- tar_data %>%
  #pad(interval = time_step, group = "site", break_above = 10) %>%
  complete_dates(group = "site", time_step = "hours")
nrow(tar_data_pad2)
table(tar_data_pad2$site)
range(tar_data_pad2$date)
```

```{r, include = FALSE}
c(365, 366)*24
tar_data_pad %>%
  filter(year(date) >= 2008) %>%
  group_by(site, year = year(date)) %>%
  summarise(N = n())
```

# Seleção de EMAs

## Critério de disponibilidade mínima de observações

Antes de determinar o período de dados de cada EMA é preciso remover as EMAs que conhenham muitos (todos) dados faltantes. Isso geralmente ocorre para EMAs inseridas recentemente no conjunto de dados. 

- Classificação das EMAs pela % de falhas

```{r, }
# ranking de falhas das EMAs
rank_falt <- tar_data_pad %>%
  group_by(site) %>%
  summarise(N = n(), 
            falt = n_NA(tair), 
            falt_perc = percent_NA(tair)) %>%
  arrange(desc(falt_perc)) %>%
  # combinando com info para saber nome das EMAs
  left_join(select(tar_info, site, name), by = "site")
rank_falt
not_informative <- rank_falt %>%
  filter(N == falt)
not_informative
```

Removendo EMAs sem dados.

```{r}
glimpse(tar_data_pad)
# dados preenchidos e informativo
tar_data_pad_inf <- tar_data_pad %>%
  filter(! site %in% not_informative$site)
glimpse(tar_data_pad_inf)
```


## Critérios baseados no período de dados

### Período de dados

```{r}
tar_periods <- tar_data_pad_inf %>%
  group_by(site) %>%
  filter(!is.na(tair)) %>%
  summarise(start = as.Date(min(date)) # data inicial
           ,end = as.Date(max(date)) # data final
           ,period = round(time_length(end - start, unit = "year"), 1))  %>% # período
            # desagrupando por site
            ungroup() %>%
  arrange(desc(period)) %>%
  # combinando com info para saber nome das EMAs
  left_join(select(tar_info, site, name, state, lon, lat), by = "site")
tar_periods
```



### Seleção de dados por período mínimo e ano inicial

As EMAs selecionadas devem ter um período de pelo menos 4 anos e baseado em trabalhos prévios devem começar em 2008.

```{r}
begin_year <- 2000
end_year <- 2016
period_thres <- 4

tar_data_sel <- tar_data_pad_inf %>%
  selectByDate(year = begin_year:end_year)
glimpse(tar_data_pad_inf)
# EMAs com pelo menos 4 anos (sites selecionados)
sites_sel <- tar_periods %>%
  filter(period >= period_thres) %>%
  select(site) %>%
  pull()
sites_sel
length(sites_sel)
# filtrar dados só com as EMAs do sites_sel
tar_data_sel <- filter(tar_data_sel, site %in% sites_sel) 
# calc tmed horária baseado nos extremos
tar_data_sel$tmed <- select(tar_data_sel, tmax, tmin) %>%
  rowMeans(., na.rm = TRUE) %>%
  ifelse(is.na(.), NA, .)
tar_data_sel <- arrange_vars(tar_data_sel, c("tmed" = 3))
glimpse(tar_data_sel)
```

Salvar dados selecionados pelos critérios de período mínimo de 4 anos e ano inicial 2008.

```{r}
# dados
file_data_sel <- paste0("tar-data-inmet-", 
                        begin_year, "-", 
                        end_year, "-",
                        period_thres, "yrs-",
                        "south.rds")
saveRDS(tar_data_sel, file = file.path("../output", file_data_sel))
# metadados
file_info_sel <- paste0("tar-info-inmet-", 
                        begin_year, "-", 
                        end_year, "-",
                        period_thres, "yrs-",
                        "south.rds")
tar_info_sel <- filter(tar_info, site %in% sites_sel)
tar_info_sel
saveRDS(tar_info_sel, file = file.path("../output", file_info_sel))
```




Gráfico do período de dados.

```{r plot-periodo, fig.width=8.3, fig.height=6.25, fig.align='center' }
estados <- readRDS("../data/estados_sul.rds")
filter(tar_periods, site == "A825")
gg_periodo <- gg_bubble(#data =  filter(tar_periods, site %in% sites_sel)
                     data = tar_periods
                     ,z = "period"
                     ,breaks = pretty(tar_periods$period, n = 10)
                     ,limites = estados
                     ,colors_z = viridis::viridis
                     ,color_fill = "burlywood3"
                     ,z_legend = "Período (anos)"
                     ,text_color = "gray30"
                     ,point_color = "transparent"
                     ,text_size = 2.6
                     ,point_size = 3
                     ,repel_min_seg_len = 2,
                     ,legend.justification = c(0,1) 
                     ,legend.position = c(0.01, 0.99)
                     ,guide_type = "legend") 
gg_periodo
```

```{r}
ggplot(tar_periods, 
       aes(x = round(period))) + 
  #geom_histogram(aes(y = ..count../sum(..count..)), stat = "count") + 
  geom_histogram(stat = "count", aes(fill = state)) +
  #viridis::scale_fill_viridis(discrete=TRUE, option = "inferno") +
  scale_fill_grey() +
  #scale_y_continuous(labels = scales::percent,
  #                   sec.axis = sec_axis(~. * sum(.))) + 
  scale_x_continuous(breaks = scales::pretty_breaks(n = 15), 
                     expand = c(0, 0)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 12), 
                     expand = c(0.05, 0.05)) +
  labs(x = "Período (anos)", y = "Nº de EMAs") +
  theme_bw()
```


# Disponibilidade de dados

```{r}
tar_gaps <- tar_data_sel %>%
  group_by(site) %>%
  dplyr::summarise(# % de dados faltantes
                   missing = round(percent_NA(tair), 1),
                   valid = 100 - missing,
                   #long_valid = length(na.contiguous(tair)),
                   long_valid = longest_gap(tair, reverse = TRUE),
                   # tamnho da falha mais longa
                   long_gap = longest_gap(tair),
                   #data de início da maior falha
                   date_slg = date_longest_gap(tair, date),
                   date_slv = date_longest_gap(tair, date, reverse = TRUE)
                   )  %>%
  # desagrupando por site
  ungroup() %>%
  # combinando com a tabela de informações das EMA
  left_join(tar_periods, by = "site") %>%
  arrange_vars(c("name" = 2, "period" = 3)) %>%
  arrange(desc(missing))
tar_gaps
```

```{r}
# convertendo max duração contínua de horas faltantes ou válidas para dias
tar_gaps <- tar_gaps %>%
  mutate(long_valid = round(time_length(hours(long_valid), unit = "day")),
         long_gap = round(time_length(hours(long_valid), unit = "day")))
tar_gaps
```


```{r}
file_gaps <- paste0("tar-gaps-inmet-", 
                        begin_year, "-", 
                        end_year, "-",
                        period_thres, "yrs-",
                        "south.rds")
saveRDS(tar_gaps, file = file.path("../output", file_gaps))
```


*Porcentagem de observações por mês *

```{r, message=FALSE, warning=FALSE}
table(tar_data_sel$site)
# porcentagem mensal de dados válidos com o periodo nao normalizado
records_month <- tar_data_sel %>%
  select(date, site, tair, prec) %>%
  group_by(site) %>%
  openair::timeAverage(avg.time = "month", 
                       statistic = "data.cap", 
                       type = "site") %>%
  #mutate(date = as.Date(date)) %>%
  ungroup() %>%
  mutate(x = as.numeric(date)) 
records_month 
```

Evolução do nº de EMAs no tempo.
```{r}
limiar <- 60
n_emas_by_month <- records_month %>% 
  mutate(is_avail = ifelse(tair >= limiar, 1, 0)) %>%
  group_by(date) %>%
  summarise(N = sum(is_avail))
n_emas_by_month

n_emas_by_month %>%
  ggplot(., aes(x = date, y = N)) +
  geom_line() +
  labs(y = "nº de EMAs") +
  scale_x_datetime(date_breaks = "12 months", 
                   date_labels = "%Y") +
  theme_bw()

```


```{r}
# ordem dos sites por disponibilidade
sites_ord <- records_month %>%
  group_by(site) %>%
  summarise(disp = mean(tair, na.rm = TRUE)) %>%
  arrange(desc(disp)) %>%
  #select(site) %>%
  pull(site) %>%  as.character()
#sites_ord
# dados para plot
records_month_plot <- records_month %>%
  mutate(tair = ifelse(tair > 0, tair, NA),
         prec = ifelse(prec > 0, prec, NA),
         site = ordered(site, levels = sites_ord))
levels(records_month_plot$site)
```




**Gráficos da disponibilidade de dados.**

```{r, fig.height=14, fig.asp=1.2}
ggp <- ggplot(aes(x = date,
                  y = site),
              data = records_month_plot) +
              #data = records_month_plot %>% filter(year(date)==2016)) +
  #geom_tile(aes(fill = avail)) +
  #geom_raster(aes(fill = avail)) + 
  geom_point(aes(colour = tair), shape=15, size = 2.6) + 
  labs(x = "Ano", y = "EMA") +
  #ylim(rev(levels(records_month_plot$site)))+
  #scale_y_discrete(expand = c(0, 0)
                   #labels = yl_labels
  #                 ) +
   #scale_y_discrete(position = c("right"),
    #                labels = yr_labels, ) +
  scale_x_datetime(expand = c(0, 0), 
                   #breaks = scales::date_breaks(width = "1 year"),
                   #minor_breaks = scales::date_breaks(width = "6 months"),
                   #labels = scales::date_format("%Y")
                   ) +
  scale_colour_gradientn("obs/month\n(%)",
                         colours = viridis::viridis(n = 256),
                         na.value = NA) +
#  scale_fill_gradientn("records/month\n(%)",
#                       colours = viridis::viridis(n= 256),
#                       na.value = NA)+ 
  theme_bw() +
  theme(text = element_text(size=10)
        #aspect.ratio = 1
        #aspect.ratio = 2 / (1 + sqrt(5))
        #aspect.ratio = (1 + sqrt(5))/ 2
        )
ggp
```



```{r plot-disp, fig.width=9.3, fig.height=7.25, fig.align='center' }
estados <- readRDS("../data/estados_sul.rds")
filter(tar_gaps, site == "A825")
gg_disp <- gg_bubble(#data =  filter(tar_gaps, site %in% sites_sel)
                     data = tar_gaps
                     ,z = "valid"
                     ,breaks = c(pretty(tar_gaps$valid, n = 10),105)
                     ,limites = estados
                     ,colors_z = viridis::viridis
                     ,color_fill = "burlywood3"
                     ,z_legend = "Disponibilidade (%)"
                     ,text_color = "gray30"
                     ,point_color = "transparent"
                     ,text_size = 2.6
                     ,point_size = 3
                     ,repel_min_seg_len = 2,
                     ,legend.justification = c(0,1) 
                     ,legend.position = c(0.01, 0.99)
                     ,guide_type = "colourbar"
                     ) 
gg_disp
range(tar_gaps$valid)
```


```{r}
ggplot(tar_gaps, 
       aes(x = round(missing))) + 
  #geom_histogram(aes(y = ..count../sum(..count..)), stat = "count") + 
  geom_histogram(stat = "count", aes(fill = state)) +
  #viridis::scale_fill_viridis(discrete=TRUE, option = "inferno") +
  scale_fill_grey() +
  #scale_y_continuous(labels = scales::percent,
  #                   sec.axis = sec_axis(~. * sum(.))) + 
  scale_y_continuous(breaks = scales::pretty_breaks(n = 10)) +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 12), 
                     expand = c(0, 0)) +
  labs(x = "Dados faltantes (%)", y = "Nº de EMAs") +
  theme_bw()
```




```{r plots1, fig.width=9.3, fig.height=7.25, fig.align='center' }
cowplot::plot_grid(gg_periodo, 
                   gg_disp,
                   labels = c("(a)", "(b)"),
                   vjust = 4.8,
                   hjust = -7.25,
                   align = "h", 
                   label_size = 20)
```

# Arquivos gerados para análises futuras


- arquivo de metadados
    
    - `output/tar-info-inmet-2008-2016-4yrs-south.rds`
    
- arquivo de dados meteorológicos

    - `output/tar-data-inmet-2008-2016-4yrs-south.rds`
